requirements_txt.txt
streamlit==1.28.0
opencv-python-headless==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
Pillow==10.0.1
streamlit-webrtc==0.47.1
av==10.0.0